feat(s3): the s3 driver support was implemented

+++ src/database/migrations/2021_01_01_000000_add_caption_field_to_media.php
@@ -13,9 +13,13 @@ class AddCaptionFieldToMedia extends Migration
      */
     public function up()
     {
-        Schema::table('media', function (Blueprint $table) {
-            $table->string('caption', 512)->nullable()->after('size');
-        });
+        if (Schema::hasTable('media')) {
+            Schema::table('media', function (Blueprint $table) {
+                if (!Schema::hasColumn('media', 'caption')) {
+                    $table->string('caption', 512)->nullable()->after('size');
+                }
+            });
+        }
     }
 
     /**
@@ -25,8 +29,10 @@ class AddCaptionFieldToMedia extends Migration
      */
     public function down()
     {
-        Schema::table('media', function (Blueprint $table) {
-            $table->dropColumn('caption');
-        });
+        if (Schema::hasTable('media')) {
+            Schema::table('media', function (Blueprint $table) {
+                $table->dropColumn('caption');
+            });
+        }
     }
 }
